- enabled placing order api connecting backend and paypal - added order and order detail components to display past orders - fixed filter product logic - fixed other breaking logic

+++ Backend/Parth Traders.Data/DataModel/Admin/ProductDataModel.cs
@@ -34,7 +34,7 @@ namespace Parth_Traders.Data.DataModel.Admin
         public ICollection<OrderDetailDataModel> OrderDetails { get; set; }
 
         [Required]
-        public int ProductRating { get; set; }
+        public double ProductRating { get; set; }
 
         [Required]
         public long UnitPrice { get; set; }
